# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VSTR Management System API",
        "description": "Visitor Management System API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Server"
        },
        {
            "url": "https://api.vstrmgmt.com",
            "description": "Production Server"
        },
        {
            "url": "https://staging-api.vstrmgmt.com",
            "description": "Staging Server"
        }
    ],
    "paths": {
...
        "/admin/locations/{location}/visit-types": {
            "get": {
                "tags": [
                    "VisitTypes"
                ],
                "summary": "List visit types (admin)",
                "description": "List visit types for a location (admin)",
                "operationId": "adminListVisitTypes",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "VisitTypes"
                ],
                "summary": "Create visit type",
                "description": "Create a visit type for a location",
                "operationId": "adminStoreVisitType",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/kiosk/locations/{location}/visit-types": {
            "get": {
                "tags": [
                    "VisitTypes"
                ],
                "summary": "List visit types (kiosk)",
                "description": "List visit types for a location (kiosk)",
                "operationId": "kioskListVisitTypes",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/locations/{location}/visit-types/{visit_type}": {
            "get": {
                "tags": [
                    "VisitTypes"
                ],
                "summary": "Show visit type (admin)",
                "description": "Retrieve a visit type",
                "operationId": "adminShowVisitType",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "visit_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "VisitTypes"
                ],
                "summary": "Update visit type",
                "description": "Update a visit type for a location",
                "operationId": "adminUpdateVisitType",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "visit_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "VisitTypes"
                ],
                "summary": "Delete visit type",
                "description": "Delete a visit type",
                "operationId": "adminDeleteVisitType",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "visit_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/kiosk/locations/{location}/visit-types/{visit_type}": {
            "get": {
                "tags": [
                    "VisitTypes"
                ],
                "summary": "Show visit type (kiosk)",
                "description": "Retrieve a visit type for kiosk clients",
                "operationId": "kioskShowVisitType",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "visit_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/locations/{location}/visit-types/{visit_type}/form-fields": {
            "post": {
                "tags": [
                    "VisitTypes"
                ],
                "summary": "Create form field for visit type",
                "description": "Create a form field for a visit type",
                "operationId": "adminStoreVisitTypeFormField",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "visit_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/kiosk/visit-responses": {
            "post": {
                "tags": [
                    "VisitResponses"
                ],
                "summary": "Store a visit response",
                "description": "Create a new visit response from kiosk",
                "operationId": "storeVisitResponse",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/kiosk/visit-responses/{visitId}": {
            "get": {
                "tags": [
                    "VisitResponses"
                ],
                "summary": "Get visit response",
                "description": "Retrieve visit response by id",
                "operationId": "getVisitResponse",
                "parameters": [
                    {
                        "name": "visitId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/visits/by-qr/{qrCode}": {
            "get": {
                "tags": [
                    "VisitResponses"
                ],
                "summary": "Get visit by QR",
                "description": "Lookup a visit by its QR code",
                "operationId": "getVisitByQr",
                "parameters": [
                    {
                        "name": "qrCode",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/kiosk/visits/by-id/{idNumber}": {
            "get": {
                "tags": [
                    "VisitResponses"
                ],
                "summary": "Get visit by ID number",
                "description": "Lookup a visit by visitor ID number",
                "operationId": "getVisitByIdNumber",
                "parameters": [
                    {
                        "name": "idNumber",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/kiosk/visits/{visitId}/checkout": {
            "patch": {
                "tags": [
                    "VisitResponses"
                ],
                "summary": "Checkout a visit",
                "description": "Marks a visit as checked out",
                "operationId": "checkoutVisit",
                "parameters": [
                    {
                        "name": "visitId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}